small fixes of delete service and added validation of user to delete

+++ app/services/User.py
@@ -10,12 +10,15 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 class UserService:
 
+
     def __init__(self, session: AsyncSession) -> None:
         self.__userRepo = UserRepo(session=session)
-    
+
+
     async def show_users(self, filters: UserFilterDTO) -> Sequence[UserORM]:
         return await self.__userRepo.get_all_users(filters)
-    
+
+
     async def add_user(self, user: UserCreateDTO) -> UserORM:
         user_with_same_email = await self.__userRepo.select_user_by_email(email=user.email)
 
@@ -23,22 +26,28 @@ class UserService:
             raise HTTPException(status_code=400, detail="User with same email exists!")
 
         return await self.__userRepo.create_user(user)
-    
+
+
     async def patch_user(self, user_id, user_schema: UserUpdate) -> UserORM:
         user_db = await self.__userRepo.select_user_by_id(user_id=user_id)
         if not user_db:
             raise HTTPException(status_code=404, detail="User not found!")
         return await self.__userRepo.update_user(user_db=user_db, user_schema=user_schema)
-    
-    async def delete_user(self, user_id: int) -> None:
+
+
+    async def delete_user(self, user_id: int, used_who_asked_id: int) -> None:
+        if user_id == used_who_asked_id:
+            raise HTTPException(status_code=403, detail="Cant delete yourself!")
         user_db = await self.__userRepo.select_user_by_id(user_id=user_id)
         if not user_db:
             raise HTTPException(status_code=404, detail="User not found!")
         await self.__userRepo.remove_user(user_db=user_db)
         return
-    
+
+
     async def login(self, user: UserLogin) -> Tokens:
         user_db = await self.__userRepo.select_user_by_email(email=user.email)
+
         if not user_db:
             raise HTTPException(status_code=400, detail="Incorrect creds!")
         if not Crypt.check_password(user.password, user_db.password_hash):
